Rename divider files to match their exports
`common.ts` was vague; `dividerStyles.ts` misled (it holds a
CSS-generating function, not static styles). Each file's new name
matches its single export.
