+++ packages/mail-react/src/components/divider/defaultDividerStyles.ts
