+++ packages/mail-react/src/components/divider/dividerProps.ts
