+++ packages/mail-react/src/components/divider/generateResponsiveDividerCss.ts
